Migrate useNotificationHistory hook to Util
