+++ .gitignore
@@ -3,4 +3,5 @@
 **/coverage
 **/.DS_Store
 .npmrc
+**/.env
