+++ packages/notifi-react-card/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './AlertConfiguration';
 export * from './DeepPartialReadonly';
+export * from './NotificationHistoryUtils';
